hackathonerds/requirements.txt
# Core ML
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.5.1
sentence-transformers>=3.0.1
lightgbm>=4.6.0

# STT
RealtimeSTT
faster-whisper>=1.0.0

# AI / LLM
openai>=1.0.0
python-dotenv>=1.0.0

# API server
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Data generation (demo/testing)
faker>=24.0.0

# STT live dependencies
rich
pyautogui
colorama
torch
torchaudio
